% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnFit.R
\name{convertBN}
\alias{convertBN}
\title{Converting TPDA structure result to bnlearn network format}
\usage{
convertBN(TPDAstruct)
}
\value{
bayesian network structure
}
\description{
Using TPDA structure result and graphNEL() to construct Network in R
}
