% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpda.R
\name{dfProcess}
\alias{dfProcess}
\title{Pre-process data to better learn Bayesian networks}
\usage{
dfProcess(df, discrete = F)
}
\arguments{
\item{df}{a data frame containing the the variables in the model}
}
\value{
a matrix containing cleand values
}
\description{
Pre-process data to better learn Bayesian networks
}
