% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpda.R
\name{getCMI}
\alias{getCMI}
\title{Calculating conditional mutual info}
\usage{
getCMI(gxi, gyi, cutset, gen)
}
\arguments{
\item{gxi}{source node}

\item{gyi}{target node}

\item{cutset}{cutoff}

\item{gen}{all genes}
}
\value{
will be saved as [from,to,value]
}
\description{
Calculating conditional mutual info
}
