% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpda.R
\name{getDoubleHang_value}
\alias{getDoubleHang_value}
\title{Calculating the value of covariance matrix}
\usage{
getDoubleHang_value(vectorA, vectorB)
}
\description{
Calculating the covariance matrix value of vector A and vector B
}
