% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpda.R
\name{getMI}
\alias{getMI}
\title{Calculating mutual info}
\usage{
getMI(df, weight)
}
\arguments{
\item{df}{the file path which you input}
}
\value{
will be saved as [from,to,value]
}
\description{
Calculating mutual info
}
