% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpda.R
\name{TPDA_algorithm}
\alias{TPDA_algorithm}
\title{Three Phase Development Algorithm}
\usage{
TPDA_algorithm(filePath, weight1, weight2, weight3, gene2MI)
}
\arguments{
\item{filePath}{a file path of gene expression data}

\item{weight1, weight2, weight3}{weight values, mutual info threshold for  phase1, phase2, phase3}

\item{gene2MI}{a data frame containing gene-gene mutual information}
}
\value{
bayesian network structure
}
\description{
Three Phase Development Algorithm
}
